requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
gunicorn==25.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pydub==0.25.1
python-dotenv==1.2.1
Werkzeug==3.1.5
yt-dlp==2026.2.4
